Revert "feat: availableTimes 파싱 적용 (#37)"
This reverts commit 7e9c6c74125bfb997d32d19fdad0dc0867d4a9d5.
