+++ app/(home)/reservation/_hooks/useReserveCoffeeChat.ts
@@ -1,17 +1,17 @@
 import useCreateCoffeeChat from "@/apis/coffeechat/hooks/useCreateCoffeeChat";
-import type { PostCoffeeChatRequest } from "@/apis/coffeechat/types";
-import type { ScheduleForm } from "@/app/(home)/schedule/types/scheduleForm";
+import { PostCoffeeChatRequest } from "@/apis/coffeechat/types";
+import { CoffeeChatForm } from "@/types/coffeechat";
 import { toYYYYMMDD } from "@/utils/dateUtils";
 
 const useReserveCoffeeChat = () => {
   const { isSuccess, mutate: createCoffeeChat } = useCreateCoffeeChat();
 
   const reserveCoffeeChat = (
-    { date, timeRange, ...coffeeChatForm }: ScheduleForm,
+    { date, time, ...coffeeChatForm }: CoffeeChatForm,
     { mentor, mentee }: Pick<PostCoffeeChatRequest, "mentor" | "mentee">
   ) => {
     const YYYYMMDD = toYYYYMMDD(date);
-    const [startTime, endTime] = timeRange.split(" ~ ");
+    const [startTime, endTime] = time.split(" ~ ");
 
     createCoffeeChat({ mentor, mentee, date: YYYYMMDD, startTime, endTime, ...coffeeChatForm });
   };
