--- app/(home)/schedule/types/scheduleForm.ts
@@ -1,12 +1,0 @@
-type FirstStep = {
-  date: Date;
-  timeRange: string;
-};
-
-type SecondStep = {
-  question: string;
-};
-
-type ScheduleForm = FirstStep & SecondStep;
-
-export type { FirstStep, ScheduleForm, SecondStep };
