+++ types/coffeechat.ts
@@ -1,5 +1,3 @@
-import { WeekType } from "@/types/user";
-
 type CoffeeChatStatus = "AGREE" | "REQUEST" | "SUGGEST" | "DONE" | "CANCEL";
 
 type AvailableTimeRange = {
@@ -7,10 +5,15 @@ type AvailableTimeRange = {
   endTime: string;
 };
 
-type AvailableTimes = Array<
-  AvailableTimeRange & {
-    week: WeekType;
-  }
->;
+type FirstStepData = {
+  date: Date;
+  time: string;
+};
 
-export type { AvailableTimeRange, AvailableTimes, CoffeeChatStatus };
+type SecondStepData = {
+  question: string;
+};
+
+type CoffeeChatForm = FirstStepData & SecondStepData;
+
+export type { AvailableTimeRange, CoffeeChatForm, CoffeeChatStatus, FirstStepData, SecondStepData };
